Moving set_dtype_dimension to tools
