+++ rsopt/libe_tools/optimizer_mobo.py
@@ -1,3 +1,4 @@
+from rsopt.libe_tools import tools
 from rsopt.libe_tools import optimizer
 from rsopt.libe_tools.generator_functions.persistent_mobo import persistent_mobo
 from libensemble.alloc_funcs.start_only_persistent import only_persistent_gens
@@ -11,7 +12,7 @@ class MoboOptimizer(optimizer.libEnsembleOptimizer):
         super().__init__()
 
     def _configure_optimizer(self):
-        gen_out = [optimizer.set_dtype_dimension(dtype, self.dimension) for dtype in mobo_gen_out]
+        gen_out = [tools.set_dtype_dimension(dtype, self.dimension) for dtype in mobo_gen_out]
         user_keys = {'lb': self.lb,
                      'ub': self.ub,
                      'dim': self._config.get_dimension(),
